rename flows and jobs, adding "Configuration"
