+++ src/main/java/uk/ac/ebi/eva/pipeline/configuration/jobs/flows/PopulationStatisticsFlowConfiguration.java
@@ -15,6 +15,7 @@
  */
 package uk.ac.ebi.eva.pipeline.configuration.jobs.flows;
 
+import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.job.builder.FlowBuilder;
 import org.springframework.batch.core.job.flow.Flow;
@@ -23,34 +24,35 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.core.task.SimpleAsyncTaskExecutor;
 
-import static uk.ac.ebi.eva.pipeline.configuration.BeanNames.CALCULATE_STATISTICS_OPTIONAL_FLOW;
-import static uk.ac.ebi.eva.pipeline.configuration.BeanNames.PARALLEL_STATISTICS_AND_ANNOTATION;
-import static uk.ac.ebi.eva.pipeline.configuration.BeanNames.VEP_ANNOTATION_OPTIONAL_FLOW;
+import uk.ac.ebi.eva.pipeline.configuration.jobs.steps.CalculateStatisticsStepConfiguration;
+import uk.ac.ebi.eva.pipeline.configuration.jobs.steps.LoadStatisticsStepConfiguration;
+
+import static uk.ac.ebi.eva.pipeline.configuration.BeanNames.CALCULATE_STATISTICS_FLOW;
+import static uk.ac.ebi.eva.pipeline.configuration.BeanNames.CALCULATE_STATISTICS_STEP;
+import static uk.ac.ebi.eva.pipeline.configuration.BeanNames.LOAD_STATISTICS_STEP;
 
 /**
- * Configuration class that defines a flow that executes in parallel the annotation and the statistics flows.
+ * Configurations that defines the calcule statistics process. First calculate the statistics then load them to
+ * the knowledge base.
  */
 @Configuration
 @EnableBatchProcessing
-@Import({AnnotationFlowOptional.class, PopulationStatisticsOptionalFlow.class})
-public class ParallelStatisticsAndAnnotationFlow {
+@Import({CalculateStatisticsStepConfiguration.class, LoadStatisticsStepConfiguration.class})
+public class PopulationStatisticsFlowConfiguration {
 
     @Autowired
-    @Qualifier(VEP_ANNOTATION_OPTIONAL_FLOW)
-    private Flow annotationFlowOptional;
+    @Qualifier(CALCULATE_STATISTICS_STEP)
+    private Step calculateStatisticsStep;
 
     @Autowired
-    @Qualifier(CALCULATE_STATISTICS_OPTIONAL_FLOW)
-    private Flow optionalStatisticsFlow;
+    @Qualifier(LOAD_STATISTICS_STEP)
+    private Step loadStatisticsStep;
 
-    @Bean(PARALLEL_STATISTICS_AND_ANNOTATION)
-    public Flow parallelStatisticsAndAnnotation() {
-        return new FlowBuilder<Flow>(PARALLEL_STATISTICS_AND_ANNOTATION)
-                .split(new SimpleAsyncTaskExecutor())
-                .add(optionalStatisticsFlow, annotationFlowOptional)
-                .build();
+    @Bean(CALCULATE_STATISTICS_FLOW)
+    public Flow calculateStatisticsOptionalFlow() {
+        return new FlowBuilder<Flow>(CALCULATE_STATISTICS_FLOW)
+                .start(calculateStatisticsStep).next(loadStatisticsStep).build();
     }
 
 }
