+++ src/main/java/uk/ac/ebi/eva/pipeline/parameters/validation/job/AnnotationJobParametersValidator.java
@@ -21,15 +21,16 @@ import org.springframework.batch.core.JobParametersValidator;
 import org.springframework.batch.core.job.CompositeJobParametersValidator;
 import org.springframework.batch.core.job.DefaultJobParametersValidator;
 
-import uk.ac.ebi.eva.pipeline.parameters.validation.step.AnnotationLoaderStepParametersValidator;
 import uk.ac.ebi.eva.pipeline.parameters.validation.step.AnnotationMetadataStepParametersValidator;
 import uk.ac.ebi.eva.pipeline.parameters.validation.step.GenerateVepAnnotationStepParametersValidator;
+import uk.ac.ebi.eva.pipeline.parameters.validation.step.LoadVepAnnotationStepParametersValidator;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Validates the job parameters necessary to execute an {@link uk.ac.ebi.eva.pipeline.jobs.AnnotationJob}
+ * Validates the job parameters necessary to execute an
+ * {@link uk.ac.ebi.eva.pipeline.configuration.jobs.AnnotationJobConfiguration}
  */
 public class AnnotationJobParametersValidator extends DefaultJobParametersValidator {
 
@@ -44,7 +45,7 @@ public class AnnotationJobParametersValidator extends DefaultJobParametersValida
         boolean studyIdRequired = false;
 
         jobParametersValidators.add(new GenerateVepAnnotationStepParametersValidator(studyIdRequired));
-        jobParametersValidators.add(new AnnotationLoaderStepParametersValidator(studyIdRequired));
+        jobParametersValidators.add(new LoadVepAnnotationStepParametersValidator(studyIdRequired));
         jobParametersValidators.add(new AnnotationMetadataStepParametersValidator());
 
         CompositeJobParametersValidator compositeJobParametersValidator = new CompositeJobParametersValidator();
