+++ src/test/java/uk/ac/ebi/eva/pipeline/configuration/jobs/steps/CreateDatabaseIndexesStepTest.java
@@ -32,7 +32,7 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import uk.ac.ebi.eva.pipeline.configuration.BeanNames;
-import uk.ac.ebi.eva.pipeline.configuration.jobs.DatabaseInitializationJob;
+import uk.ac.ebi.eva.pipeline.configuration.jobs.DatabaseInitializationJobConfiguration;
 import uk.ac.ebi.eva.test.configuration.BatchTestConfiguration;
 import uk.ac.ebi.eva.test.rules.TemporaryMongoRule;
 import uk.ac.ebi.eva.test.utils.JobTestUtils;
@@ -47,7 +47,7 @@ import static uk.ac.ebi.eva.test.utils.JobTestUtils.assertCompleted;
  */
 @RunWith(SpringRunner.class)
 @TestPropertySource({"classpath:common-configuration.properties", "classpath:test-mongo.properties"})
-@ContextConfiguration(classes = {DatabaseInitializationJob.class, BatchTestConfiguration.class})
+@ContextConfiguration(classes = {DatabaseInitializationJobConfiguration.class, BatchTestConfiguration.class})
 public class CreateDatabaseIndexesStepTest {
 
     private static final String COLLECTION_FEATURES_NAME = "features";
